+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # tpopup change log
 
+## v1.1 2026-05-17
+Added the grammar correction mode `-grammar`
+
 ## v1.0 2026-05-17
-Version 1.0, an app that works only with the -translate mode.
+Version 1.0, an app that works only with the `-translate` mode.